requirements.txt
Django>=4.2.8,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2 >=2.9.3,<2.10
drf-spectacular>=0.27.0,<0.28
Pillow>=10.1.0,<10.2
uwsgi>=2.0.23,<2.1
djangorestframework_simplejwt>=5.3.1,<5.4

pytest>=7.4.4,<7.5
pytest-django>=4.7.0,<4.8
factory_boy>=3.3.0,<3.4
Faker>=22.0.0,<22.1

djangorestframework-simplejwt>=5.3.1,<5.4